Refactoring
Removed old stuff

+++ project/GameProject/Game/Components/BaseComponent.hpp
@@ -9,7 +9,7 @@ class BaseComponent
 {
 public:
 
-	virtual uint64_t getComponentID() const = 0;
+	virtual uint32_t getComponentID() const = 0;
 	void setAttachedOn(GameObject* owner);
 private:
