+++ project/GameProject/Game/Components/RenderComponent.cpp
@@ -1,15 +1,15 @@
 #include "RenderComponent.hpp"
 
-const uint64_t IBaseComponent<RenderComponent>::typeID = (uint64_t)Components::RENDER_COMPONENT;
+const uint32_t IBaseComponent<RenderComponent>::typeID = (uint32_t)Components::RENDER_COMPONENT;
 
 void RenderComponent::init() {
 	instanceId = 0;
 }
 
-void RenderComponent::setInstanceId(uint64_t _instanceId) {
+void RenderComponent::setInstanceId(uint32_t _instanceId) {
 	instanceId = _instanceId;
 }
 
-uint64_t RenderComponent::getInstanceId() const {
+uint32_t RenderComponent::getInstanceId() const {
 	return instanceId;
 }
